adds media picker to AddBikeScreen.kt, adds city picker and price validation + correct keyboard type for price.

+++ app/src/main/java/com/bikerental/app/data/model/Bike.kt
@@ -10,7 +10,7 @@ data class Bike(
 
     val bikeName: String = "",
 
-    val price: Int = 0,
+    val price: Double = 0.00,
 
     val imageUrl: String = "",
